fix(PPDSC-2039): move eventHandler to composable
